services/diffusers-tryon/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
torch>=2.2.0
Pillow>=10.0.0
